correcciones pre-entrega

+++ src/router/products.router.js
@@ -9,9 +9,9 @@ router.get("/", async (req, res) => {
     try {
         const limit = parseInt(req.query?.limit);
         const products = await productManager.getProducts(limit);
-        res.json(products);
+        res.status(200).json({ products });
     } catch (error) {
-        res.status(500).send({ error: "Error al obtener los productos" });
+        res.status(500).send({ error: error.message });
     };
 });
 
@@ -19,9 +19,10 @@ router.get("/:pid", async (req, res) => {
     try {
         const pid = parseInt(req.params.pid);
         const product = await productManager.getProductByID(pid);
-        res.json(product);
+        if (!product) return res.status(404).json({ error: `El producto con ID ${pid} no existe` });
+        return res.status(200).json({ product });
     } catch (error) {
-        res.status(404).send({ error: "Product not found" });
+        res.status(500).send({ error: error.message });
     };
 });
 
@@ -29,9 +30,10 @@ router.post("/", async (req, res) => {
     try {
         const data = req.body;
         const product = await productManager.addProduct(data);
-        res.json(product);
+        if (product) return res.status(201).json({ message: "Producto agregado con éxito" });
+        return res.status(404).json({ error: "Error al agregar el producto" });
     } catch (error) {
-        res.status(500).send({ error: "Error al agregar el producto" });
+        res.status(500).send({ error: error.message });
     };
 });
 
@@ -39,20 +41,24 @@ router.put("/:pid", async (req, res) => {
     try {
         const pid = parseInt(req.params.pid);
         const data = req.body;
-        const product = await productManager.updateProduct(pid, data);
-        res.json(product); 
+        const producto = await productManager.getProductByID(pid);
+        if (!producto) return res.status(404).json({ error: `El producto con ID ${pid} no existe` });
+        await productManager.updateProduct(pid, data);
+        res.status(200).json({ message: `Producto ${pid} actualizado con éxito` });
     } catch (error) {
-        res.status(500).send({ error: "Error al actualizar el producto" });
+        res.status(500).send({ error: error.message });
     };
 });
 
 router.delete("/:pid", async (req, res) => {
     try {
         const pid = parseInt(req.params.pid);
-        const product = await productManager.deleteProduct(pid);
-        res.json(product);
+        const producto = await productManager.getProductByID(pid);
+        if (!producto) return res.status(404).json({ error: `El producto con ID ${pid} no existe` });
+        await productManager.deleteProduct(pid);
+        res.status(200).json({ message: `Producto ${pid} eliminado con éxito` });
     } catch (error) {
-        res.status(500).send({ error: "Error al eliminar el producto" });
+        res.status(500).send({ error: error.message });
     }
 });
